Move login and signup form in ui folder
